shimexpy_gpu/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shimexpy-gpu"
version = "0.1.0"
description = "GPU acceleration module for shimexpy, utilizing CuPy for high-performance spatial harmonics analysis."
authors = [{name = "Jorge Luis Beltran Diaz"}]
license = { text = "Apache-2.0" }
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Environment :: GPU :: NVIDIA CUDA",
]
dependencies = [
    "shimexpy>=0.1.0"
]

[project.optional-dependencies]
# Users should select the appropriate CuPy package for their CUDA version
cuda13x = ["cupy-cuda13x"]
cuda12x = ["cupy-cuda12x"]
cuda11x = ["cupy-cuda11x"]

[tool.setuptools.packages.find]
where = ["."]
include = ["shimexpy_gpu*"]
